support "modfs boot-manager" and "7412 volume counter" patches also in "no freetz"-mode [1]
[1] https://freetz.org/wiki/help/howtos/development/repack_fw#Verwendungvonfwmodimnofreetz-Modus


git-svn-id: http://svn.freetz.org/trunk@14823 f5190166-0702-4917-9039-51ec32eddaf5

+++ patches/scripts/800-modfs_boot_manager.sh
@@ -1,15 +1,20 @@
 [ "$FREETZ_PATCH_MODFS_BOOT_MANAGER" = "y" ] || return 0;
 
-echo1 "adding modfs boot-manager"
-for step in precheck install postcheck; do
-	TARGET_BRANDINGS=all \
-	TARGET_BRANDING=all \
-	source "${TOOLS_DIR}/modfs/modscripts/gui_boot_manager_v0.4" \
-		en \
-		"${FILESYSTEM_MOD_DIR}" \
-		"MODE (UNUSED)" \
-		"${step}"
-	[ $? -eq 0 ] || error 1 "adding modfs boot-manager failed in step \"${step}\""
+for oem in $(supported_brandings) all; do
+	www_oem="${FILESYSTEM_MOD_DIR}/usr/www/${oem}"
+	[ -d "${www_oem}" -a ! -L "${www_oem}" ] || continue
+
+	echo1 "adding modfs boot-manager to branding \"${oem}\""
+	for step in precheck install postcheck; do
+		TARGET_BRANDINGS=${oem} \
+		TARGET_BRANDING=${oem} \
+		source "${TOOLS_DIR}/modfs/modscripts/gui_boot_manager_v0.4" \
+			en \
+			"${FILESYSTEM_MOD_DIR}" \
+			"MODE (UNUSED)" \
+			"${step}"
+		[ $? -eq 0 ] || error 1 "adding modfs boot-manager failed in step \"${step}\""
+	done
 done
 
 # replace modfs version of boot-manager with a more recent one from YourFritz project
